+++ CHANGELOG.md
@@ -29,6 +29,7 @@ with the `YY.MINOR.MICRO` scheme.
 - Supported backoff strategies when configuring vshard from the governor.
 - Fixed restart the whole cluster at once taking too long.
 - Fixed an issue where `cluster.shredding` option does not get applied on instance restart.
+- `EXPLAIN` queries now support `OPTION` (e.g. `EXPLAIN SELECT 1 OPTION (SQL_VDBE_OPCODE_MAX = 6)`)
 
 ## [25.3.1] - 2025-07-25